Refactors strategies
